refactor: types and search result mapping
