+++ packages/frontend/src/features/omnibar/components/Omnibar.tsx
@@ -1,6 +1,6 @@
 import { getSearchResultId, SearchItemType } from '@lightdash/common';
 import { Input, Loader, MantineProvider, Modal, Stack } from '@mantine/core';
-import { useDisclosure, useHotkeys } from '@mantine/hooks';
+import { useDebouncedValue, useDisclosure, useHotkeys } from '@mantine/hooks';
 import { IconSearch } from '@tabler/icons-react';
 import { FC, MouseEventHandler, useState } from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
@@ -9,9 +9,9 @@ import { useProject } from '../../../hooks/useProject';
 import { useValidationUserAbility } from '../../../hooks/validation/useValidation';
 import { useTracking } from '../../../providers/TrackingProvider';
 import { EventName } from '../../../types/Events';
-import { useDebouncedSearch } from '../hooks/useDebouncedSearch';
-import { OMNIBAR_MIN_QUERY_LENGTH } from '../hooks/useSearch';
+import useSearch, { OMNIBAR_MIN_QUERY_LENGTH } from '../hooks/useSearch';
 import { allSearchItemTypes, SearchItem } from '../types/searchItem';
+import { isSearchResultEmpty } from '../utils/isSearchResultEmpty';
 import OmnibarEmptyState from './OmnibarEmptyState';
 import OmnibarFilters from './OmnibarFilters';
 import OmnibarItemGroups from './OmnibarItemGroups';
@@ -30,11 +30,17 @@ const Omnibar: FC<Props> = ({ projectUuid }) => {
     const canUserManageValidation = useValidationUserAbility(projectUuid);
 
     const [query, setQuery] = useState<string>();
+    const [debouncedValue] = useDebouncedValue(query, 300);
+
+    const { data: searchResults, isFetching } = useSearch(
+        projectUuid,
+        debouncedValue,
+    );
 
     const [isOmnibarOpen, { open: openOmnibar, close: closeOmnibar }] =
         useDisclosure(false);
 
-    const handleSpotlightOpenInputClick: MouseEventHandler<HTMLInputElement> = (
+    const handleOmnibarOpenInputClick: MouseEventHandler<HTMLInputElement> = (
         e,
     ) => {
         e.currentTarget.blur();
@@ -49,7 +55,7 @@ const Omnibar: FC<Props> = ({ projectUuid }) => {
         openOmnibar();
     };
 
-    const handleSpotlightOpenHotkey = () => {
+    const handleOmnibarOpenHotkey = () => {
         track({
             name: EventName.GLOBAL_SEARCH_OPEN,
             properties: {
@@ -61,10 +67,10 @@ const Omnibar: FC<Props> = ({ projectUuid }) => {
     };
 
     useHotkeys([
-        ['mod + k', handleSpotlightOpenHotkey, { preventDefault: true }],
+        ['mod + k', handleOmnibarOpenHotkey, { preventDefault: true }],
     ]);
 
-    const handleSpotlightClose = () => {
+    const handleOmnibarClose = () => {
         track({
             name: EventName.GLOBAL_SEARCH_CLOSED,
             properties: {
@@ -103,15 +109,13 @@ const Omnibar: FC<Props> = ({ projectUuid }) => {
         }
     };
 
-    const { items, isFetching } = useDebouncedSearch(projectUuid, query);
-
     const [openPanels, setOpenPanels] =
         useState<SearchItemType[]>(allSearchItemTypes);
 
     return (
         <>
             {!isOmnibarOpen && (
-                <OmnibarTarget onOpen={handleSpotlightOpenInputClick} />
+                <OmnibarTarget onOpen={handleOmnibarOpenInputClick} />
             )}
 
             <MantineProvider inherit theme={{ colorScheme: 'light' }}>
@@ -122,7 +126,7 @@ const Omnibar: FC<Props> = ({ projectUuid }) => {
                     closeOnClickOutside
                     closeOnEscape
                     opened={isOmnibarOpen}
-                    onClose={handleSpotlightClose}
+                    onClose={handleOmnibarClose}
                     yOffset={100}
                     styles={{
                         body: {
@@ -174,16 +178,16 @@ const Omnibar: FC<Props> = ({ projectUuid }) => {
                                     projectData?.name ?? 'your project'
                                 }.`}
                             />
-                        ) : isFetching ? (
+                        ) : !searchResults ? (
                             <OmnibarEmptyState message="Searching..." />
-                        ) : items.length === 0 ? (
+                        ) : isSearchResultEmpty(searchResults) ? (
                             <OmnibarEmptyState message="No results found." />
                         ) : (
                             <>
                                 <OmnibarFilters />
 
                                 <OmnibarItemGroups
-                                    items={items}
+                                    searchResults={searchResults}
                                     projectUuid={projectUuid}
                                     canUserManageValidation={
                                         canUserManageValidation
