+++ packages/frontend/src/features/omnibar/components/OmnibarItemGroups.tsx
@@ -2,43 +2,37 @@ import { SearchItemType } from '@lightdash/common';
 import { Accordion, Text } from '@mantine/core';
 import { FC, useMemo } from 'react';
 import { SearchItem } from '../types/searchItem';
+import { SearchResultMap } from '../types/searchResultMap';
 import { getSearchItemLabel } from '../utils/getSearchItemLabel';
 import OmnibarItem from './OmnibarItem';
 
 type Props = {
     openPanels: SearchItemType[];
     onOpenPanelsChange: (panels: SearchItemType[]) => void;
-    items: SearchItem[];
+    searchResults: SearchResultMap;
     projectUuid: string;
     canUserManageValidation: boolean;
     onClick: (item: SearchItem) => void;
 };
 
-type GroupedItems = Partial<Record<SearchItemType, SearchItem[]>>;
-
 const OmnibarItemGroups: FC<Props> = ({
     openPanels,
     onOpenPanelsChange,
     projectUuid,
-    items,
+    searchResults,
     canUserManageValidation,
     onClick,
 }) => {
-    const itemsGroupedByType = useMemo(() => {
-        return items.reduce<GroupedItems>((acc, item) => {
-            return { ...acc, [item.type]: (acc[item.type] ?? []).concat(item) };
-        }, {});
-    }, [items]);
+    console.log(searchResults);
 
     const sortedGroupEntries = useMemo(() => {
-        const entries = Object.entries(itemsGroupedByType) as Array<
-            [SearchItemType, Array<SearchItem>]
-        >;
-
-        return entries.sort(([_typeA, itemsA], [_typeB, ItemsB]) => {
-            return (itemsA[0].searchRank ?? 0) - (ItemsB[0].searchRank ?? 0);
-        });
-    }, [itemsGroupedByType]);
+        return Object.entries(searchResults)
+            .filter(([_type, items]) => items.length > 0)
+            .sort(
+                ([_typeA, itemsA], [_typeB, ItemsB]) =>
+                    (itemsA[0].searchRank ?? 0) - (ItemsB[0].searchRank ?? 0),
+            );
+    }, [searchResults]);
 
     return (
         <Accordion
@@ -65,11 +59,11 @@ const OmnibarItemGroups: FC<Props> = ({
                 onOpenPanelsChange(newPanels)
             }
         >
-            {sortedGroupEntries.map(([type, groupItems]) => (
-                <Accordion.Item key={type} value={type}>
+            {sortedGroupEntries.map(([groupType, groupItems]) => (
+                <Accordion.Item key={groupType} value={groupItems[0].type}>
                     <Accordion.Control>
                         <Text color="dark" fw={500} fz="xs">
-                            {getSearchItemLabel(type)}
+                            {getSearchItemLabel(groupItems[0].type)}
                         </Text>
                     </Accordion.Control>
