+++ packages/frontend/src/features/omnibar/hooks/useSearch.ts
@@ -1,6 +1,8 @@
 import { ApiError, SearchFilters, SearchResults } from '@lightdash/common';
 import { useQuery } from '@tanstack/react-query';
 import { getSearchResults } from '../api/search';
+import { SearchResultMap } from '../types/searchResultMap';
+import { getSearchItemMap } from '../utils/getSearchItemMap';
 
 export const OMNIBAR_MIN_QUERY_LENGTH = 3;
 
@@ -8,13 +10,13 @@ const useSearch = (
     projectUuid: string,
     query: string = '',
     filters?: SearchFilters,
-) => {
-    return useQuery<SearchResults, ApiError>({
+) =>
+    useQuery<SearchResults, ApiError, SearchResultMap>({
         queryKey: ['search', query],
         queryFn: () => getSearchResults({ projectUuid, query, filters }),
         retry: false,
         enabled: query.length >= OMNIBAR_MIN_QUERY_LENGTH,
+        select: (data) => getSearchItemMap(data, projectUuid),
     });
-};
 
 export default useSearch;
