+++ packages/frontend/src/features/omnibar/types/searchResultMap.ts
@@ -1,0 +1,3 @@
+import { getSearchItemMap } from '../utils/getSearchItemMap';
+
+export type SearchResultMap = ReturnType<typeof getSearchItemMap>;
