+++ packages/frontend/src/features/omnibar/utils/isSearchResultEmpty.ts
@@ -1,0 +1,5 @@
+import { SearchResultMap } from '../types/searchResultMap';
+
+export const isSearchResultEmpty = (searchResult: SearchResultMap) => {
+    return Object.values(searchResult).every((value) => value.length === 0);
+};
